end of mobile development
due to request by  Grace Community Church mobile development will here by cease on this branch. that is the branch of this project that connects to  Adventure Club. other future branches that that do not in any way relate to Adventure Club. may be build as a different type of quiz game base solely on the bible.